Fixed paths after linter->reducers move
